@@ -3,6 +3,7 @@
   "version": "2.1.1",
   "description": "pull iconfont, download iconfont",
   "main": "./dist/index.js",
+  "type": "module",
   "bin": {
     "pull-iconfont": "./dist/index.js"
   },
@@ -11,7 +12,7 @@
     "dist"
   ],
   "scripts": {
-    "test": "./dist/index.js -c ./example/local/.pulliconfontrc-2.js",
+    "test": "./dist/index.js -c ./example/.pulliconfontrc.js",
     "build": "tsc --project tsconfig.json",
     "lint:ts": "eslint ./src --fix --ext .ts,.tsx"
   },
@@ -34,19 +35,18 @@
   "dependencies": {
     "@babel/parser": "^7.16.4",
     "chalk": "^2.4.2",
-    "commander": "^2.20.3",
+    "commander": "^12.1.0",
     "node-fetch": "^3.3.2",
     "rimraf": "^3.0.2",
     "unzipper": "^0.12.3"
   },
   "devDependencies": {
     "@hansin/lint-config": "^0.0.1",
-    "@types/commander": "^2.12.2",
     "@types/node": "^12.20.37",
     "@types/rimraf": "^3.0.2",
     "@types/unzipper": "^0.10.4",
     "husky": "^7.0.4",
     "lint-staged": "^12.1.2",
-    "typescript": "^4.5.2"
+    "typescript": "^5.5.4"
   }
 }
