@@ -1,12 +1,12 @@
-module.exports = {
-  cookie: '网站cookie拷贝到这里',
-  spm: 'url参数拷贝到这里',
-  pid: 'url参数拷贝到这里',
-  ctoken: 'url参数拷贝到这里',
+export default {
+  cookie: "网站cookie拷贝到这里",
+  spm: "url参数拷贝到这里",
+  pid: "url参数拷贝到这里",
+  ctoken: "url参数拷贝到这里",
   saveDemoFile: false,
-}
+};
 
-// module.exports = {
+// export default {
 //   downloadUrl: 'https://www.iconfont.cn/api/project/download.zip?xxxxx=xxxx',
 //   cookie: 'cna=xxxxxxxx',
 //   saveDemoFile: false, // 是否保存demo文件
