@@ -1,13 +1,13 @@
 import fs, { promises as fsPromises } from "fs";
-import { getOutPath } from "./getOutPath";
-import { showLog, showErrorLog, Config } from "./utils";
+import { getOutPath } from "./getOutPath.js";
+import { showLog, Config } from "./utils.js";
 import fetch from "node-fetch";
 import { join } from "path";
 import unzipper from "unzipper";
 import rimraf from "rimraf";
 const { stat, mkdir, readdir, copyFile } = fsPromises;
 
-const fileExists = async (p: string) => !!(await stat(p).catch((e) => false));
+const fileExists = async (p: string) => !!(await stat(p).catch(() => false));
 
 const download = async function (config: Config) {
   const outPath = getOutPath(config);
@@ -79,8 +79,8 @@ const download = async function (config: Config) {
 
           // 删除过渡文件
           try {
-            await rimraf.sync(join(outPath, fontDir));
-            await rimraf.sync(`${outPath}/download.zip`);
+            rimraf.sync(join(outPath, fontDir));
+            rimraf.sync(`${outPath}/download.zip`);
           } catch (err) {
             console.log(err);
           }
