@@ -1,10 +1,8 @@
-import { resolve } from 'path'
-import { Config } from './utils'
+import { resolve } from "path";
+import { Config } from "./utils.js";
 
 function getOutPath(config: Config): string {
-  return resolve(config.outputPath)
+  return resolve(config.outputPath);
 }
 
-export {
-  getOutPath,
-}
+export { getOutPath };
